Add test for mux issue when producer is always ready
When producer is always ready and ctrl interface offers data before the data
interfaces, ctrl will be consumed without waiting for the data. This happens
because the following verilog process is generated (with some of the names
substituted) for driving the ready signals:

    // Comb block for: inputs
    always_comb begin
        ctrl.ready = 0;
        din0.ready = 0;
        din1.ready = 0;
        if (ctrl.valid) begin
            ctrl.ready = dout.ready;
            if (ctrl_s == 0) begin
                din0.ready = din.ready;
            end
            if (ctrl_s == 1) begin
                din1.ready = din.ready;
            end
        end
    end

Issue arrises because ctrl.ready does not depend on whether the appropriate mux
channel has valid data.